Fix signed unsigned conversion

+++ src/CryptoNoteCore/Currency.cpp
@@ -490,8 +490,8 @@ difficulty_type Currency::nextDifficulty(uint8_t blockMajorVersion, std::vector<
 		int64_t L(0);
 		uint64_t next_D;
 
-		for (int64_t i = 1; i <= N; i++) {
-			L += std::max<int64_t>(-7 * T, std::min<int64_t>(7 * T, timestamps[i] - timestamps[i - 1])) * i;
+		for (size_t i = 1; i <= N; i++) {
+			L += (int64_t)(std::max<int64_t>(-7 * T, std::min<int64_t>(7 * T, timestamps[i] - timestamps[i - 1])) * i);
 		}
 		if (L < 1)
 			L = 1;
